requirements.txt
# Core PDF processing
PyMuPDF>=1.23.0

# NLP and text processing
spacy>=3.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Vector database and RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM integration
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Regular expressions and pattern matching
regex>=2023.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
loguru>=0.7.0

# Optional: Web interface
# gradio>=4.0.0
